Adds custom hook to make logic re-usable and cleaner code
